+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.4.0 (Unreleased)
 **Bug Fixes**
 - [#1426](https://github.com/Azure/azure-storage-fuse/issues/1426) Read panic in block-cache due to boundary conditions.
+- Do not allow mount path and temp-cache path to be same when using block-cache.
+- Do not allow to mount with non-empty directory provided for disk persistence in block-cache.
 - Rename file was calling an additional getProperties call.
 
 **Features**
@@ -8,6 +10,9 @@
 
 **Other Changes**
 - Stream config will be converted to block-cache config implicitly and 'stream' component is no longer used from this release onwards.
+
+**Other Changes**
+- MSI login with object-id will not rely on azcli anymore.
 
 ## 2.3.2 (2024-09-03)
 **Bug Fixes**